ST: do cluster operator setup correctly  (#2080)
* Set default operation timeout to default value, improve CO log assertations and add option for better test debug


* fixup! Set default operation timeout to default value, improve CO log assertations and add option for better test debug


* fixup! fixup! Set default operation timeout to default value, improve CO log assertations and add option for better test debug

+++ test/src/main/java/io/strimzi/test/BaseITST.java
@@ -157,7 +157,7 @@ public class BaseITST {
         bindingsNamespaces = namespaces;
         for (String namespace: namespaces) {
 
-            if (kubeClient().getNamespace(namespace) != null) {
+            if (kubeClient().getNamespace(namespace) != null && System.getenv("SKIP_TEARDOWN") == null) {
                 LOGGER.warn("Namespace {} is already created, going to delete it", namespace);
                 kubeClient().deleteNamespace(namespace);
                 cmdKubeClient().waitForResourceDeletion("Namespace", namespace);
